Fix INT8 attention default in launcher, suite, vram_budget, stress_test
v0.1.40 fixed the attention_quant default to bf16 in config.py and
server.py but missed 5 other locations that still defaulted to int8.
INT8 attention causes catastrophic perplexity (~5600 vs ~10.6) and
garbage output. This was the likely cause of garbage on WSL.

Fixed: launcher.py, suite.py (2 places), vram_budget.py, stress_test.py

IMPORTANT: Users with an existing ~/.krasis/config from before this fix
will have CFG_ATTENTION_QUANT=int8 saved. Delete ~/.krasis/config or
manually change to bf16.

+++ python/krasis/launcher.py
@@ -376,7 +376,7 @@ class LauncherConfig:
         self.kv_dtype: str = "fp8_e4m3"
         self.gpu_expert_bits: int = 4
         self.cpu_expert_bits: int = 4
-        self.attention_quant: str = "int8"
+        self.attention_quant: str = "bf16"
         self.shared_expert_quant: str = "int8"
         self.dense_mlp_quant: str = "int8"
         self.lm_head_quant: str = "int8"
